requirement.txt
tensorflow
Pillow
streamlit